+++ node/package.json
@@ -1,0 +1,7 @@
+{
+  "private": true,
+  "main": "../dist/es5/index-node.js",
+  "jsnext:main": "../dist/es2015/index-node.js",
+  "module": "../dist/es2015/index-node.js",
+  "types": "../dist/es2015/index-node.d.ts"
+}
