+++ src/index-node.ts
@@ -1,0 +1,1 @@
+export { discoverProjectStyles } from './style-discovery';
