+++ src/index.ts
@@ -4,18 +4,21 @@ import { getCriticalRules, extractCriticalRules, getCriticalStyles, getUsedStyle
 import { alterProjectStyles } from './operations';
 import { createCriticalStyleStream } from './reporters/critical';
 import { createStyleStream } from './reporters/used';
-import { discoverProjectStyles, loadStyleDefinitions, parseProjectStyles } from './scanForStyles';
-
+import { loadStyleDefinitions, parseProjectStyles } from './style-operations';
 import { createUsedFilter as createUsedSelectorsFilter } from './utils/cache';
 
+/**
+ * @deprecated please import discoverProjectStyles from 'used-styles/node'
+ */
+export const discoverProjectStyles = () => {
+  throw new Error("Please import discoverProjectStyles from 'used-styles/node'");
+};
+
 export { serializeStylesLookup, loadSerializedLookup } from './serialize';
-
 export { UsedTypes, StyleDefinition, SelectionFilter } from './types';
-
 export {
   createUsedSelectorsFilter,
   loadStyleDefinitions,
-  discoverProjectStyles,
   parseProjectStyles,
   alterProjectStyles,
   getUsedStyles,
