+++ tsconfig.json
@@ -12,7 +12,7 @@
     "noImplicitReturns": true,
     "noFallthroughCasesInSwitch": true,
     "noImplicitAny": true,
-    "removeComments": true,
+    "removeComments": false,
     "importHelpers": true,
     "target": "es6",
     "moduleResolution": "node",
